+++ apps/aspiration/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from apps.aspiration.views.aspirations import AspirationUpdateDestroyView, AspirationProgressDetailUpdateDestroyView, AspirationListCreateView, AspirationDetailView, AspirationProgressListCreateView
+from apps.aspiration.views.aspirations import AspirationStatsView, AspirationUpdateDestroyView, AspirationProgressDetailUpdateDestroyView, AspirationListCreateView, AspirationDetailView, AspirationProgressListCreateView
 from apps.aspiration.views.categories import CategoryListView, CategoryCreateView, CategoryDetailUpdateDestroyView
 from apps.aspiration.views.notifications import NotificationListView, NotificationReadView, NotificationDetailDestroyView    
 
@@ -13,6 +13,7 @@ urlpatterns = [
     path('aspiration/list/<uuid:pk>/manage/', AspirationUpdateDestroyView.as_view(), name='aspiration-manage'),
     path('aspiration/progress/', AspirationProgressListCreateView.as_view(), name='progress-list-create'),
     path('aspiration/progress/<uuid:pk>/manage/', AspirationProgressDetailUpdateDestroyView.as_view(), name='progress-detail-manage'),
+    path('aspiration/stats/', AspirationStatsView.as_view(), name='aspiration-stats'),
     
     path('aspiration/notifications/', NotificationListView.as_view(), name='notification-list'),
     path('aspiration/notifications/<uuid:pk>/read/', NotificationReadView.as_view(), name='notification-read'),